conf/requirements.txt
amqp==5.2.0
asgiref==3.7.2
async-timeout==4.0.3
billiard==4.2.0
celery==5.3.6
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
Django==3.2.25
django-filter==23.5
djangorestframework==3.12.4
djangorestframework-simplejwt==5.3.1
flower==2.0.1
humanize==4.9.0
kombu==5.3.5
pillow==10.2.0
prometheus_client==0.20.0
prompt-toolkit==3.0.43
PyJWT==2.8.0
python-dateutil==2.9.0.post0
pytz==2024.1
redis==5.0.3
six==1.16.0
sqlparse==0.4.4
tornado==6.4
typing_extensions==4.10.0
tzdata==2024.1
vine==5.1.0
wcwidth==0.2.13
